+++ packages/base-ui/registry/index.ts
@@ -9,11 +9,6 @@ const dir = path.join(path.dirname(fileURLToPath(import.meta.url)), '../src');
 export const registry: Registry = {
   name: 'fumadocs/base-ui',
   dir,
-  tsconfigPath: '../tsconfig.json',
-  packageJson: '../package.json',
-  env: {
-    ui: '@fumadocs/base-ui',
-  },
   components: [
     ...commonComponents,
     ...(await findSlotComponents(dir)),
@@ -46,11 +41,17 @@ export const registry: Registry = {
           type: 'layout',
           path: 'layouts/shared/index.tsx',
           target: '<dir>/shared/index.tsx',
+          meta: {
+            preserveAs: 'layouts/shared',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/shared/client.tsx',
           target: '<dir>/shared/client.tsx',
+          meta: {
+            preserveAs: 'layouts/shared',
+          },
         },
       ],
     },
@@ -61,16 +62,25 @@ export const registry: Registry = {
           type: 'layout',
           path: 'layouts/docs/index.tsx',
           target: '<dir>/docs/index.tsx',
+          meta: {
+            preserveAs: 'layouts/docs',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/docs/client.tsx',
           target: '<dir>/docs/client.tsx',
+          meta: {
+            preserveAs: 'layouts/docs',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/docs/page/index.tsx',
           target: '<dir>/docs/page/index.tsx',
+          meta: {
+            preserveAs: 'layouts/docs/page',
+          },
         },
       ],
       unlisted: true,
@@ -82,11 +92,17 @@ export const registry: Registry = {
           type: 'layout',
           path: 'layouts/flux/index.tsx',
           target: '<dir>/flux/index.tsx',
+          meta: {
+            preserveAs: 'layouts/flux',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/flux/page/index.tsx',
           target: '<dir>/flux/page/index.tsx',
+          meta: {
+            preserveAs: 'layouts/flux/page',
+          },
         },
       ],
       unlisted: true,
@@ -98,16 +114,25 @@ export const registry: Registry = {
           type: 'layout',
           path: 'layouts/notebook/index.tsx',
           target: '<dir>/notebook/index.tsx',
+          meta: {
+            preserveAs: 'layouts/notebook',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/notebook/client.tsx',
           target: '<dir>/notebook/client.tsx',
+          meta: {
+            preserveAs: 'layouts/notebook',
+          },
         },
         {
           type: 'layout',
           path: 'layouts/notebook/page/index.tsx',
           target: '<dir>/notebook/page/index.tsx',
+          meta: {
+            preserveAs: 'layouts/notebook/page',
+          },
         },
       ],
       unlisted: true,
@@ -119,6 +144,9 @@ export const registry: Registry = {
           type: 'layout',
           path: 'layouts/home/index.tsx',
           target: '<dir>/home/index.tsx',
+          meta: {
+            preserveAs: 'layouts/home',
+          },
         },
         {
           type: 'ui',
@@ -300,7 +328,6 @@ export const registry: Registry = {
     },
     {
       name: 'ai/page-actions',
-      title: 'Page Actions',
       description: 'Common page actions',
       files: [
         {