+++ packages/openapi/tsdown.config.ts
@@ -1,5 +1,7 @@
 import { defineConfig } from 'tsdown';
-import { compileInline } from './scripts/compile-inline.utils';
+import { mkdir, writeFile } from 'node:fs/promises';
+import path from 'node:path';
+import { Scanner } from '@tailwindcss/oxide';
 
 export default defineConfig({
   format: 'esm',
@@ -15,7 +17,6 @@ export default defineConfig({
     './src/server/index.ts',
     './src/requests/generators/*.ts',
   ],
-  fixedExtension: false,
   unbundle: true,
   dts: {
     sourcemap: false,
@@ -44,6 +45,30 @@ export default defineConfig({
       v['./css/*'] = './css/*';
       return v;
     },
-    legacy: true,
   },
 });
+
+async function compileInline() {
+  await mkdir('css/generated', { recursive: true });
+  const scanner = new Scanner({
+    sources: [
+      {
+        base: path.resolve('src'),
+        pattern: '{playground,scalar,ui}/**/*.{ts,tsx}',
+        negated: false,
+      },
+      {
+        base: path.resolve('src'),
+        pattern: 'server/**/*.tsx',
+        negated: false,
+      },
+    ],
+  });
+  await writeFile('css/generated/shared.css', namesToFile(scanner.scan()));
+
+  console.log('generated CSS files');
+}
+
+function namesToFile(names: string[]) {
+  return names.map((name) => `@source inline(${JSON.stringify(name)});`).join('\n');
+}