+++ packages/radix-ui/src/layouts/docs/page/slots/toc.tsx
@@ -57,7 +57,7 @@ export function TOC({ container, header, footer, style = 'normal', list }: TOCPr
       id="nd-toc"
       {...container}
       className={cn(
-        'sticky top-(--fd-docs-row-1) h-[calc(var(--fd-docs-height)-var(--fd-docs-row-1))] flex flex-col [grid-area:toc] w-(--fd-toc-width) pt-12 pe-4 pb-2 max-xl:hidden',
+        'sticky top-(--fd-docs-row-1) h-[calc(var(--fd-docs-height)-var(--fd-docs-row-1))] flex flex-col [grid-area:toc] w-(--fd-toc-width) pt-12 pe-4 pb-2 xl:layout:[--fd-toc-width:268px] max-xl:hidden',
         container?.className,
       )}
     >